requirements.txt
awesome-slugify==1.6.5
Django==1.9.6
django-mysql==1.0.9
gunicorn==19.6.0
mysqlclient==1.3.7
Pillow==3.2.0
pytz==2016.4
regex==2016.6.2
six==1.10.0
Unidecode==0.4.19
whitenoise==3.2
